JWT + everything w/o Redis

+++ src/main/java/com/app/todoapp/entity/UserTeam.java
@@ -1,10 +1,11 @@
 package com.app.todoapp.entity;
 
 import lombok.Data;
+import org.springframework.context.annotation.Bean;
 
 @Data
 public class UserTeam {
-    private int userId;
-    private int teamId;
+    private Long userId;
+    private Long teamId;
     private String role;
 }
